api/requirements.txt
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
python-dotenv==0.17.0
SQLAlchemy==1.4.9
uWSGI==2.0.19.1
Werkzeug==1.0.1
